+++ autosubmit/config/basicconfig.py
@@ -37,7 +37,7 @@ class BasicConfig:
                 pr[name] = value
         return pr
 
-    DB_DIR = os.path.join(os.path.expanduser('~'), 'debug', 'autosubmit')
+    DB_DIR = str(Path.home() / 'debug' / 'autosubmit')
     STRUCTURES_DIR = os.path.join(
         '/esarchive', 'autosubmit', 'as_metadata', 'structures')
     GLOBAL_LOG_DIR = os.path.join(
@@ -49,7 +49,8 @@ class BasicConfig:
     AUTOSUBMIT_API_URL = "http://192.168.11.91:8081"
     DB_FILE = 'autosubmit.db'
     AS_TIMES_DB = 'as_times.db'
-    DB_PATH = os.path.join(DB_DIR, DB_FILE)
+    DB_PATH = str(Path(DB_DIR, DB_FILE))
+    AS_TIMES_DB_PATH = str(Path(DB_DIR, AS_TIMES_DB))
     LOCAL_ROOT_DIR = DB_DIR
     LOCAL_TMP_DIR = 'tmp'
     LOCAL_ASLOG_DIR = 'ASLOGS'
@@ -88,9 +89,8 @@ class BasicConfig:
         """
         Updates commonly used composed paths
         """
-        # Just one needed for the moment.
-        BasicConfig.DB_PATH = os.path.join(
-            BasicConfig.DB_DIR, BasicConfig.DB_FILE)
+        BasicConfig.DB_PATH = str(Path(BasicConfig.DB_DIR, BasicConfig.DB_FILE))
+        BasicConfig.AS_TIMES_DB_PATH = str(Path(BasicConfig.DB_DIR, BasicConfig.AS_TIMES_DB))
 
     @staticmethod
     def __read_file_config(file_path):